Create results.php
This file calculates the results and displays the answer given with the correct answer that was stored in the database. The answers are both first converted to lowercase to remove any case-sensitive errors that could result. The user's score is also shown at the bottom. The user then has the option to store their name and view the high scores.